Allow uploadFileToFolder to take a file-like object
Prior to this, uploadFileToFolder could only be called from the same
machine it was running on since it took a local filepath. This is
similar to how uploadFileToItem operates, however we should steer away
from this as we want the girder-cli library to be usable from remote
worker instances.

uploadFileToItem has been in a girder-client release and as such,
can't be fixed to conform to its folder counterpart without breaking
backwards compatibility.